requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
boto3==1.35.38
botocore==1.35.38
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
setuptools==75.6.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.2
orjson==3.10.7
pandas==2.2.3
paramiko==3.5.0
pfzy==0.3.4
prettytable==3.11.0
prompt_toolkit==3.0.48
propcache==0.2.0
py-cpuinfo==9.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
runpod==1.7.2
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.38.6
tabulate==0.9.0
threadpoolctl==3.5.0
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.66.5
tqdm-loggable==0.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.14.0
